MAF-POC-MCP/requirements.txt
fastapi
uvicorn
pydantic
agent-framework-core==1.0.0b251120
python-dotenv
beartype
azure-cosmos==4.3.0
six==1.16.0
azure-search-documents>=11.7.0b1
tenacity==8.2.3
tiktoken==0.12.0
pdfplumber
logger
azure-ai-documentintelligence
ijson
agent-framework-devui==1.0.0b251120
azure-storage-blob
agent-framework-ag-ui
fastmcp==2.13.2
aiohttp